+++ src/main/java/prefolio/prefolioserver/service/JwtTokenService.java
@@ -7,13 +7,18 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
+import prefolio.prefolioserver.domain.User;
+import prefolio.prefolioserver.error.CustomException;
 import prefolio.prefolioserver.error.ErrorCode;
+import prefolio.prefolioserver.repository.UserRepository;
 
 import java.nio.charset.StandardCharsets;
-import java.util.Base64;
-import java.util.Date;
+import java.util.*;
+
+import static prefolio.prefolioserver.error.ErrorCode.USER_NOT_FOUND;
 
 @Slf4j
 @Service
@@ -21,6 +26,8 @@ import java.util.Date;
 public class JwtTokenService {
 
     private final UserDetailsServiceImpl userDetailsService;
+    private final UserRepository userRepository;
+    private final JwtTokenService jwtTokenService;
     @Value("${jwt.secret}")
     private String JWT_SECRET;
 
@@ -56,14 +63,6 @@ public class JwtTokenService {
         return false;
     }
 
-
-    public Authentication getAuthentication(String token) {
-        UserDetails userDetails = userDetailsService.loadUserByUsername(String.valueOf(
-                this.getUserIdFromJwtToken(token)));
-        return new UsernamePasswordAuthenticationToken(userDetails, "", userDetails.getAuthorities());
-    }
-
-
     public Long getUserIdFromJwtToken(String token) {
         Claims claims = Jwts.parser()
                 .setSigningKey(Base64.getEncoder().encodeToString(("" + JWT_SECRET).getBytes(
@@ -71,5 +70,80 @@ public class JwtTokenService {
                 .parseClaimsJws(token)
                 .getBody();
         return Long.parseLong(claims.getSubject());
+    }
+
+    // JWT 토큰 리턴
+    public List<String> usersAuthorizationInput(User user) {
+        UserDetailsImpl userDetails = userDetailsService.loadUserByUsername(String.valueOf(user.getId()));
+        String accessToken = generateJwtAccessToken(userDetails);
+        String refreshToken = generateJwtRefreshToken(userDetails);
+        user.setRefreshToken(refreshToken);
+        userRepository.save(user);
+
+        List<String> token = new ArrayList<>();
+        token.add(accessToken);
+        token.add(refreshToken);
+
+        return token;
+    }
+
+//    public Authentication getAuthentication(String token) {
+//        UserDetails userDetails = userDetailsService.loadUserByUsername(String.valueOf(
+//                this.getUserIdFromJwtToken(token)));
+//        return new UsernamePasswordAuthenticationToken(userDetails, "", userDetails.getAuthorities());
+//    }
+
+    // 로그인 처리
+    public Authentication getAuthentication(String token) {
+        UserDetailsImpl userDetails = userDetailsService.loadUserByUsername(String.valueOf(
+                getUserIdFromJwtToken(token)));
+        Authentication authentication = new UsernamePasswordAuthenticationToken(
+                userDetails,
+                "",
+                userDetails.getAuthorities()
+        );
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+        return authentication;
+    }
+
+
+    public String generateJwtAccessToken(UserDetailsImpl userDetails) {
+        User user = userRepository.findByEmail(userDetails.getEmail())
+                .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
+        Calendar cal = Calendar.getInstance();
+        cal.add(Calendar.HOUR, 3);  // 만료시간 1시간
+
+        final Date issuedAt = new Date();
+        final Date accessTokenExpiresIn = new Date(cal.getTimeInMillis());
+
+        return buildAccessToken(user.getId(), issuedAt, accessTokenExpiresIn);
+    }
+
+    public String generateJwtRefreshToken(UserDetailsImpl userDetails) {
+        User user = userRepository.findByEmail(userDetails.getEmail())
+                .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
+        Calendar cal = Calendar.getInstance();
+        cal.add(Calendar.DATE, 14);  // 만료일 14일
+
+        final Date issuedAt = new Date();
+        final Date accessTokenExpiresIn = new Date(cal.getTimeInMillis());
+
+        return buildAccessToken(user.getId(), issuedAt, accessTokenExpiresIn);
+    }
+
+    public String buildAccessToken(Long id, Date issuedAt, Date accessTokenExpiresIn) {
+        return Jwts.builder()
+                .setHeaderParam(Header.TYPE, Header.JWT_TYPE)
+                .setSubject(id.toString())
+                .setIssuer("prefolio")
+                .setIssuedAt(issuedAt)
+                .setExpiration(accessTokenExpiresIn)
+                .claim("id", id)
+                .claim("roles", "USER")
+                .signWith(
+                        SignatureAlgorithm.HS256,
+                        Base64.getEncoder().encodeToString(("" + JWT_SECRET).getBytes(StandardCharsets.UTF_8))
+                )
+                .compact();
     }
 }
